sign up form test prototype
i fucking love the ease of use with puppeteer. its quite a natural and confident feeling to test like this. still have to do a few things to work out the flow like adding vitest to replace console.logging and i need to create some sort of mechinism to increment the user info each time i run a test because it's causing errors in the server and weird errors in the test suite