move utils & db into pkg subdir
i did them both at the same time because initially i thought relative pathing was going to be more confusing than it is, as it turns out it really isn't that bad. I was thinking I would have to work relative to where i define the cwd but as it turns out, i think it is smart enough to know the cwd is the main.go, maybe that happens in nodejs too but i feel like its harder in node. i wonder what the path will be like when i move the app entry (main.go) to a sub dir as well. e.g. ("../..")