reconnect the receipt upload feature
one thing i love about what's happening here is that no matter what I change, each time i go to test the receipt upload feature it just works. This is something that would just be destroyed with each code refactor in React or the framework of your choice. Initially I was going to credit HTMX, then I got to thinking maybe it's the MVC pattern but when it comes down to I think its the no build that is really doing the shining here. That's a shoutout to HTMX I'm sure