! refactor the upload api route cause it was too big for me
we had a struct named `Receipt` that collides with a different struct in models directory. one is from the db, and one is from the raw format from the api handler. it made the most sense to rename this `Receipt` to `RawReceipt` so that refactor is in here too. As far as regression testing, it worked for me.... no but really we ont have tests set up here yet cause i havent gotten around to it.