allow for manual adding of receipts
was able to make use of the partial `ReceiptManualEdit` template instead of having to hand roll the logic for the form twice. Its tricky because the dynamic nature of the items list. I don't think the solution that was implemented is the greatest but its been a few weeks since i last saw it