dockerize app and make it prod ready
i decided to skip some features and push for mvp faster than initially anticipated. this commit wipes out that build-step because it was too complex to dockerize so I just stuffed the css that would have gone into that build step; directly into a css file in the react app. So technically theres a growing mess of css building up but I figure we'll be full tailwind soon so f it