make jtw based user auth work
oh ya know it looked like it was working before but turns out there was no persistance in the UI, although the cookie was set; i just didnt have any way to check it in the client given security restrictions in the cookie. This led to the check auth endpoint being created. Also, there was no utility on the backend to validate the JWT AND expire time considerations for the cookie. It was basically a few missing pieces that I didnt initially catch ,

TODO Puppeteer testing