secure pw reset
meaning that if the user is logged in, they can reset their pw relatively easily. I also added an enhancement to the password input criteria to limit the max length of characters on the server. so nobody uploads the bee movie as a password. thanks for the tip Wes https://www.youtube.com/watch?v=oOFrLNItiX8