fix gnarly ass state
After building the MVP, i see now that the `chatStore.sessions` is the only item i need to keep in the store as far as tracking sessions goes. The modification methods can stay, but essentially the fact that `currentSessionId`, and `currentSession` exist in state makes the app incredibly hard to reason about. We have everything we need in the `sessions` array; albeit we have to find the current session every time; but even that is okay to me becase we have the search query param to work with as the source of thruth for the current session.